% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\encoding{UTF-8}
\name{pvInd-class}
\alias{$,pvInd-method}
\alias{$<-,pvInd-method}
\alias{getAe}
\alias{getAe,pvInd-method}
\alias{getAeMargin}
\alias{getAeMargin,pvInd-method}
\alias{getCov}
\alias{getCov,pvInd-method}
\alias{getDrug}
\alias{getDrug,pvInd-method}
\alias{getDrugMargin}
\alias{getDrugMargin,pvInd-method}
\alias{naRm}
\alias{naRm,pvInd-method}
\alias{names,pvInd-method}
\alias{pvInd-class}
\alias{rbind}
\alias{rbind.pvInd}
\alias{show,pvInd-method}
\title{Formal class "pvInd" and basic associated functions}
\description{
The class pvInd is a class (S4) for storing individual spontaneous report data. A number of basic functions are provided to manipulate and access elements stored in this object
}
\section{Slots}{

\describe{
\item{\code{drug}}{Drug matrix in sparse format \code{Matrix}.}

\item{\code{ae}}{Adverse Event matrix in sparse format \code{Matrix}.}

\item{\code{drugMargin}}{Vector of the marginal counts for each drug.}

\item{\code{aeMargin}}{Vector of the marginal counts for each ae.}

\item{\code{cov}}{\code{data.frame} containing individual covariates.}
}}
\section{Methods}{

\describe{
 \item{names}{\code{signature(x = "pvInd")}: returns the names of the slots of the pvInd object.}
 \item{show}{\code{signature(x = "pvInd")}: prints a description of the pvInd object.}
 \item{$}{\code{signature(x = "pvInd")}: similar to the  @ operator; used to access the content of slots of the object.}
 \item{$<-}{\code{signature(x = "pvInd")}: similar to the @ operator; used to replace the content of slots of the object.}
 \item{getDrug}{\code{signature(x = "pvInd")}: returns the drug matrix. This is equivalent to \code{object$drug}.}
 \item{getAe}{\code{signature(x = "pvInd")}: returns the adverse event matrix. This is equivalent to \code{object$ae}}
 \item{getDrugMargin}{\code{signature(x = "pvInd")}: returns the drug margin counts. This is equivalent to \code{object$drugMargin}}
 \item{getAeMargin}{\code{signature(x = "pvInd")}: returns the adverse event margin counts. This is equivalent to \code{object$aeMargin}}
 \item{getCov}{\code{signature(x = "pvInd")}: returns the covariate data.frame. This is equivalent to \code{object$cov}}
 \item{naRm}{\code{signature(x = "pvInd")}: remove spontaneous reports with NA values in the covariate slot.}
 \item{\code{rbind(...)}}{Take a sequence of pvInd objects and combine them by rows into a new pvInd object. Be careful as the function assumes that the covariates are the same for all pvInd objects. Also the function does not work if the ae slots or drug slots contain empty rows ; empty rows can result from the use of the \code{pvIndResize} function.}
}
}
\author{
Youness Ergaibi & Ismaïl Ahmed
}

