% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\encoding{UTF-8}
\name{pvInd}
\alias{pvInd}
\alias{pvInd,data.frame,data.frame,missing,data.frame-method}
\alias{pvInd,data.frame,data.frame,missing,missing-method}
\alias{pvInd,missing,missing,data.frame,data.frame-method}
\alias{pvInd,missing,missing,data.frame,missing-method}
\title{pvInd constructor}
\usage{
pvInd(drug, ae, adr, cov, ...)
}
\arguments{
\item{adr}{A data.frame with three columns:
\itemize{
\item spontaneous report identifier
\item drug label
\item adverse event label
}}

\item{drug}{A data.frame or a matrix with two columns
\itemize{
\item spontaneous report identifier
\item drug label
}}

\item{ae}{A data.frame or a matrix with two columns
\itemize{
\item spontaneous report identifier
\item adverse event label
}}

\item{cov}{A data.frame which contains individual covariates. It is assumed that the first column of the data.frame corresponds to the spontaneous report identifier.}

\item{...}{not of any use}
}
\value{
an object of class pvInd
}
\description{
This function (which takes the same name as its corresponding formal class \code{\link{pvInd-class}}) is to be used to convert spontaneous reports data to a pvInd object.
}
\author{
Ismaïl Ahmed
}
\seealso{
\code{\link{pvInd-class}}
}
\keyword{pvInd}

