% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ADRminer}
\alias{ADRminer}
\alias{ADRminer-package}
\title{ADRminer: Adverse Drug event Reporting systems miner}
\description{
\code{ADRminer} is dedicated to the automated generation of drug safety signals from spontaneous reporting databases. While currently under development, this package will likely replace the PhViD package. Great efforts were made to re-think the structure of the objects manipulated. In particular, ADRminer handles individual spontaneous reports and thus makes it  possible to conduct stratified analyses according to several covariates. From a more technical point of view, objects manipulated (\code{\link{pvInd}} and \code{\link{pvCont}}) are of class S4 and rely on the sparse matrix representation (Matrix package) in order to save as much as random access memory as possible.
}
\details{
For the frequentist methods, the package requires the LBE procedure that is stored in the Bioconductor website \url{http://bioconductor.org/}. LBE can be installed by entering \code{source("http://bioconductor.org/biocLite.R")} \code{biocLite("LBE")} in the R console.
}

