\docType{class}
\encoding{UTF-8}
\name{pvInd-class}
\alias{names,pvInd-method}
\alias{pvInd-class}
\alias{shows,pvInd-method}
\title{pvInd class}
\description{
  pvInd class
}
\section{Methods}{
  \describe{ \item{names}{\code{signature(x = "pvInd")}:
  returns the names of the slots of the object.} }
}
\author{
  Youness Ergaibi & Ismaïl Ahmed
}

