% Generated by roxygen2 (4.0.1): do not edit by hand
\encoding{UTF-8}
\name{pvIndResize}
\alias{pvIndResize}
\title{pvIndResize}
\usage{
pvIndResize(pvIndObj, drugId = NULL, aeId = NULL, aeMarginMin = 1,
  drugMarginMin = 1)
}
\arguments{
\item{pvIndObj}{an object of class "pvInd" (see \code{\link{pvInd-class}} and \code{\link{pvInd}})}

\item{drugId}{can be a vector of characters containing drug labels to be selected or a vector of column indices or the drug matrice (object$drug)}

\item{aeId}{can be a vector of characters containing ae labels to be selected or a vector of column indices or the ae matrice (object$ae)}

\item{aeMarginMin}{Minimum number of reports in which an ae must be involved in order to be kept}

\item{drugMarginMin}{Minimum number of reports in which a drug must be involved in order to be kept}
}
\value{
an object of class pvInd
}
\description{
\code{pvIndResize} makes it possible to select sets of drugs and/or aes. It also allows to remove drug(s) and/or ae(s) whose margin are less
than a given number.
}
\author{
Ismaïl Ahmed
}

