% Generated by roxygen2 (4.0.1): do not edit by hand
\encoding{UTF-8}
\name{readFAERS}
\alias{readFAERS}
\title{read data from FAERS}
\usage{
readFAERS(drugFile, reacFile, demoFile, ROLE_COD = c("PS", "SS"),
  OCCP_COD = c("MD", "OT", "PH"), aeMarginMin = 5, drugMarginMin = 5)
}
\arguments{
\item{drugFile}{path to the DRUG***.txt ASCII file}

\item{reacFile}{path to the REAC***.txt ASCII file}

\item{demoFile}{path to the DEMO***.txt ASCII file}

\item{ROLE_COD}{a character vector corresponding to the drug's reported role in event. Only reports for drug}

\item{OCCP_COD}{a character vector corresponding to abbreviation for the reporter's type of occupation
\itemize{
\item MD: Physician
\item PH: Pharmacist
\item OT: Other health-professional
\item LW: Lawyer
\item CN: Consumer
}
The default value is set to c("MD","OT", "PH").}

\item{aeMarginMin}{Minimum number of reports in which an ae must be involved in order to be kept}

\item{drugMarginMin}{Minimum number of reports in which a drug must be involved in order to be kept}
}
\description{
\code{readFAERS} makes it possible to import and convert data from the AERS into a pvInd object. Data are available on the website of the FDA. At present, we only consider initial report (I_F_COD==I). We also keep only data for which drugs do not correspond to a valid trade name (VAL_VBM==1)
}
\author{
Ismaïl Ahmed
}

