% Generated by roxygen2 (4.0.1): do not edit by hand
\encoding{UTF-8}
\name{readFAERS}
\alias{readFAERS}
\title{read data from FAERS}
\usage{
readFAERS(drugFile, reacFile, demoFile, ROLE_COD = c("PS", "SS"),
  OCCP_COD = c("MD", "OT", "PH"))
}
\arguments{
\item{drugFile}{path to the DRUG***.txt ASCII file}

\item{reacFile}{path to the REAC***.txt ASCII file}

\item{demoFile}{path to the DEMO***.txt ASCII file}

\item{ROLE_COD}{a character vector corresponding to the drug's reported role in event. Only reports for drug with the reported role are kept in the pvInd object.
\itemize{
\item PS: Primary Suspect Drug
\item SS: Secondary Suspect Drug
\item C: Concomitant
\item I: Interacting
}}

\item{OCCP_COD}{a character vector corresponding to abbreviation for the reporter's type of occupation. Only reports for drug with the reported role are kept in the pvInd object.
\itemize{
\item MD: Physician
\item PH: Pharmacist
\item OT: Other health-professional
\item LW: Lawyer
\item CN: Consumer
}
The default value is set to c("MD","OT", "PH").}
}
\description{
\code{readFAERS} makes it possible to import and convert data from the AERS into a pvInd object. Data are available on the website of the FDA. At present, we only consider initial report (I_F_COD==I). We also keep only data for which drugs do not correspond to a valid trade name (VAL_VBM==1). For more details, please refer to the documentation provided with the FAERS files (ASC_NTS.Doc). First, we try to read the data with the \code{fread} function of the \code{data.table} package. If the latter doesn't work, the data are read with the \code{read.table} function. At present, the cleaning process for the age covariate is a bit drastic: all value < 0 are considered as NA, value associated with missing AGE_COD are considered as NA and ages entered in second are also considered as missing. Then the age value is converted into year.
}
\author{
Ismaïl Ahmed
}

